requirements.txt
Pillow>=10.0.0
aiohttp>=3.8.0
